Issue/copy checking topics section within kafka UI application (#2925)
* Commit

* Moved and renamed classes from Tests

* Added new methods and refactored getRandomMessage()

* Moved SmokeTests from tests to com.provectus.kafka.ui

* Moved SmokeTests from com.provectus.kafka.ui to Suite

* resolve conversation

* resolve conversation
