+++ kafka-ui-e2e-checks/src/main/java/com/provectus/kafka/ui/pages/TopPanel.java
@@ -1,13 +1,12 @@
 package com.provectus.kafka.ui.pages;
 
-import com.codeborne.selenide.SelenideElement;
+import static com.codeborne.selenide.Selenide.$x;
 
+import com.codeborne.selenide.SelenideElement;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.codeborne.selenide.Selenide.$x;
-
-public class TopPanel {
+public class TopPanel extends BasePage{
     protected SelenideElement kafkaLogo = $x("//a[contains(text(),'UI for Apache Kafka')]");
     protected SelenideElement kafkaVersion = $x("//a[@title='Current commit']");
     protected SelenideElement logOutBtn = $x("//button[contains(text(),'Log out')]");
