+++ kafka-ui-e2e-checks/src/main/java/com/provectus/kafka/ui/pages/topic/TopicsList.java
@@ -7,6 +7,11 @@ import com.codeborne.selenide.SelenideElement;
 import com.provectus.kafka.ui.pages.BasePage;
 import com.provectus.kafka.ui.utilities.WaitUtils;
 import io.qameta.allure.Step;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import lombok.experimental.ExtensionMethod;
 
 @ExtensionMethod(WaitUtils.class)
@@ -14,6 +19,10 @@ public class TopicsList extends BasePage {
 
     protected SelenideElement topicListHeader = $x("//h1[text()='Topics']");
     protected SelenideElement addTopicBtn = $x("//button[normalize-space(text()) ='Add a Topic']");
+    protected SelenideElement searchField = $x("//input[@placeholder='Search by Topic Name']");
+    protected SelenideElement showInternalRadioBtn = $x("//input[@name='ShowInternalTopics']");
+    protected String сolumnHeaderLocator = "//table//tr/th/div[text()='%s']";
+    protected String actionButtonLocator = "//button[text()='%s']";
 
     @Step
     public TopicsList waitUntilScreenReady() {
@@ -38,5 +47,38 @@ public class TopicsList extends BasePage {
     public TopicsList openTopic(String topicName) {
         getTableElement(topicName).shouldBe(Condition.enabled).click();
         return this;
+    }
+
+    private List<SelenideElement> getActionButtons() {
+      return Stream.of("Delete selected topics", "Copy selected topic", "Purge messages of selected topics")
+          .map(name -> $x(String.format(actionButtonLocator, name)))
+          .collect(Collectors.toList());
+    }
+
+    private List<SelenideElement> getVisibleColumnHeaders() {
+      return Stream.of("Replication Factor","Number of messages","Topic Name", "Partitions", "Out of sync replicas", "Size")
+          .map(name -> $x(String.format(сolumnHeaderLocator, name)))
+        .collect(Collectors.toList());
+    }
+
+    private List<SelenideElement> getEnabledColumnHeaders(){
+      return Stream.of("Topic Name", "Partitions", "Out of sync replicas", "Size")
+          .map(name -> $x(String.format(сolumnHeaderLocator, name)))
+          .collect(Collectors.toList());
+    }
+
+    @Step
+    public List<SelenideElement> getAllVisibleElements() {
+      List<SelenideElement> visibleElements = new ArrayList<>(getVisibleColumnHeaders());
+      visibleElements.addAll(Arrays.asList(searchField, addTopicBtn, tableGrid));
+      visibleElements.addAll(getActionButtons());
+      return visibleElements;
+    }
+
+    @Step
+    public List<SelenideElement> getAllEnabledElements() {
+      List<SelenideElement> enabledElements = new ArrayList<>(getEnabledColumnHeaders());
+      enabledElements.addAll(Arrays.asList(searchField, showInternalRadioBtn,addTopicBtn));
+      return enabledElements;
     }
 }
